move get_model
