+++ lightning_pose/models/__init__.py
@@ -1,9 +1,7 @@
 """Pose estimation model classes, re-exported at the package level."""
 
-from typing import Union
-
-from omegaconf import ListConfig
-
+from lightning_pose.models.base import check_if_semi_supervised
+from lightning_pose.models.factory import get_model
 from lightning_pose.models.heatmap_tracker import (
     HeatmapTracker,
     SemiSupervisedHeatmapTracker,
@@ -22,33 +20,14 @@ from lightning_pose.models.regression_tracker import (
 )
 
 # reassign module to make classes appear to belong here
-HeatmapTracker.__module__ = "lightning_pose.models"
-SemiSupervisedHeatmapTracker.__module__ = "lightning_pose.models"
-HeatmapTrackerMHCRNN.__module__ = "lightning_pose.models"
-SemiSupervisedHeatmapTrackerMHCRNN.__module__ = "lightning_pose.models"
-HeatmapTrackerMultiviewTransformer.__module__ = "lightning_pose.models"
-SemiSupervisedHeatmapTrackerMultiviewTransformer.__module__ = "lightning_pose.models"
-RegressionTracker.__module__ = "lightning_pose.models"
-SemiSupervisedRegressionTracker.__module__ = "lightning_pose.models"
-
-def check_if_semi_supervised(losses_to_use: ListConfig | list | None = None) -> bool:
-    """Determine from the losses config whether the model is semi-supervised.
-
-    Args:
-        losses_to_use: the cfg entry specifying unsupervised losses to use.
-
-    Returns:
-        True if the model is semi-supervised, False otherwise.
-
-    """
-    if losses_to_use is None:
-        return False
-    if len(losses_to_use) == 0:
-        return False
-    if len(losses_to_use) == 1 and losses_to_use[0] == '':
-        return False
-    return True
-
+HeatmapTracker.__module__ = 'lightning_pose.models'
+SemiSupervisedHeatmapTracker.__module__ = 'lightning_pose.models'
+HeatmapTrackerMHCRNN.__module__ = 'lightning_pose.models'
+SemiSupervisedHeatmapTrackerMHCRNN.__module__ = 'lightning_pose.models'
+HeatmapTrackerMultiviewTransformer.__module__ = 'lightning_pose.models'
+SemiSupervisedHeatmapTrackerMultiviewTransformer.__module__ = 'lightning_pose.models'
+RegressionTracker.__module__ = 'lightning_pose.models'
+SemiSupervisedRegressionTracker.__module__ = 'lightning_pose.models'
 
 ALLOWED_MODELS = (
     HeatmapTracker
@@ -63,13 +42,14 @@ ALLOWED_MODELS = (
 
 # to ignore imports for sphix-autoapidoc
 __all__ = [
-    "check_if_semi_supervised",
-    "HeatmapTracker",
-    "SemiSupervisedHeatmapTracker",
-    "HeatmapTrackerMHCRNN",
-    "SemiSupervisedHeatmapTrackerMHCRNN",
-    "HeatmapTrackerMultiviewTransformer",
-    "SemiSupervisedHeatmapTrackerMultiviewTransformer",
-    "RegressionTracker",
-    "SemiSupervisedRegressionTracker",
+    'check_if_semi_supervised',
+    'get_model',
+    'HeatmapTracker',
+    'SemiSupervisedHeatmapTracker',
+    'HeatmapTrackerMHCRNN',
+    'SemiSupervisedHeatmapTrackerMHCRNN',
+    'HeatmapTrackerMultiviewTransformer',
+    'SemiSupervisedHeatmapTrackerMultiviewTransformer',
+    'RegressionTracker',
+    'SemiSupervisedRegressionTracker',
 ]
