+++ tests/conftest.py
@@ -30,13 +30,13 @@ from lightning_pose.data.datasets import (
     MultiviewHeatmapDataset,
 )
 from lightning_pose.losses import get_loss_factories
+from lightning_pose.models import get_model
 from lightning_pose.utils.io import get_videos_in_dir
 from lightning_pose.utils.predictions import PredictionHandler
 from lightning_pose.utils.scripts import (
     get_data_module,
     get_dataset,
     get_imgaug_transform,
-    get_model,
 )
 
 TOY_DATA_ROOT_DIR = str(lp.LP_ROOT_PATH / "data" / "mirror-mouse-example")
